fix: task 1 PathType class move to entry file
